C++11: Build now compiles without errors
Due to std::complex<T>::imag() and ::real() changing their signatures, in C++11,
code dealing with these was updated to compile under both versions of C++.

+++ src/Params/PADnoteParameters.cpp
@@ -477,7 +477,7 @@ void PADnoteParameters::generatespectrum_bandwidthMode(float *spectrum,
                 float idfreq = i / (float)profilesize - 0.5f;
                 idfreq *= ibw;
                 int   spfreq  = (int) (idfreq + ibasefreq);
-                float fspfreq = fmod((double)idfreq + ibasefreq, 1.0f);
+                float fspfreq = fmodf((float)idfreq + ibasefreq, 1.0f);
                 if(spfreq <= 0)
                     continue;
                 if(spfreq >= size - 1)
